Refactored so that the modal is its own container. #8
